Move FormBuilder utils to lib
