+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/Edit.tsx
@@ -10,9 +10,9 @@ import { RefsProvider } from "./RefsContext";
 import { RichTextLocked } from "./elements";
 import { ExpandingInput } from "app/(gcforms)/[locale]/(form administration)/form-builder/components/shared";
 import { useTemplateStore } from "@lib/store";
-import { getQuestionNumber, sortByLayout } from "@clientComponents/form-builder/util";
+import { getQuestionNumber, sortByLayout } from "@lib/utils/form-builder";
 import { SettingsPanel } from "./settings/SettingsPanel";
-import { cleanInput } from "@clientComponents/form-builder/util";
+import { cleanInput } from "@lib/utils/form-builder";
 import { SaveButton } from "app/(gcforms)/[locale]/(form administration)/form-builder/components/shared/SaveButton";
 import { useRehydrate } from "@lib/hooks";
