+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/ModalRules.tsx
@@ -6,7 +6,7 @@ import { Modal, ModalButton, ModalFormRules } from "./index";
 import { FormElementWithIndex } from "@lib/types/form-builder-types";
 import { useTemplateStore, useModalRulesStore } from "@lib/store";
 import { Button } from "@clientComponents/globals";
-import { getPathString } from "@formbuilder/getPath";
+import { getPathString } from "@lib/utils/form-builder/getPath";
 import {
   ChoiceRule,
   choiceRulesToConditonalRules,
