+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/settings/SettingsDialog.tsx
@@ -21,7 +21,7 @@ import { LocalizedFormProperties } from "@lib/types/form-builder-types";
 import { ResponseEmail } from "app/(gcforms)/[locale]/(form administration)/form-builder/components/ResponseEmail";
 import { isValidGovEmail } from "@lib/validation";
 import { useTemplateApi } from "@lib/hooks";
-import { completeEmailAddressRegex } from "@clientComponents/form-builder/util";
+import { completeEmailAddressRegex } from "@lib/utils/form-builder";
 
 enum DeliveryOption {
   vault = "vault",
