+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/translate/components/DownloadCSV.tsx
@@ -5,7 +5,7 @@ import { useTemplateStore } from "@lib/store/useTemplateStore";
 import { Button } from "@clientComponents/globals";
 import { FormElement } from "@lib/types";
 import { getDate, slugify } from "@lib/client/clientHelpers";
-import { alphabet } from "@clientComponents/form-builder/util";
+import { alphabet } from "@lib/utils/form-builder";
 
 const formatText = (str: string) => `"${str}"`;
