+++ app/(gcforms)/[locale]/(form administration)/form-builder/components/ResponseEmail.tsx
@@ -3,7 +3,7 @@ import React, { useEffect } from "react";
 import { useTranslation } from "@i18n/client";
 import { isValidGovEmail } from "@lib/validation";
 import { Input } from "./shared";
-import { completeEmailAddressRegex } from "@clientComponents/form-builder/util";
+import { completeEmailAddressRegex } from "@lib/utils/form-builder";
 
 const HintText = ({ id, children }: { id: string; children?: JSX.Element | string }) => {
   return (
