+++ app/(gcforms)/[locale]/(form administration)/form-builder/components/navigation/LinksSubMenu.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from "@i18n/client";
 import * as DropdownMenuPrimitive from "@radix-ui/react-dropdown-menu";
 import { useTemplateStore } from "@lib/store/useTemplateStore";
 import { ShareExternalLinkIcon, CopyIcon } from "@serverComponents/icons";
-import { getHost } from "@clientComponents/form-builder/util";
+import { getHost } from "@lib/utils/form-builder";
 
 interface LinkItem {
   name: string;
