--- components/clientComponents/form-builder/example-form.json
@@ -1,246 +1,0 @@
-{
-  "titleEn": "Lemonade Stand Funding",
-  "titleFr": "[fr] Lemonade Stand Funding",
-  "introduction": {
-    "descriptionEn": "Introduction",
-    "descriptionFr": "[FR] Introduction"
-  },
-  "privacyPolicy": {
-    "descriptionEn": "Privacy statement",
-    "descriptionFr": "Déclaration de confidentialité"
-  },
-  "confirmation": {
-    "descriptionEn": "#Your submission has been received \n\r Thank you for your interest. We will contact you within 3-5 business days.  \n\r <a href='https://digital.canada.ca'>Go back</a>",
-    "descriptionFr": "#[fr] Your submission has been received. \n\r Thank you for your interest. We will contact you within 3-5 business days. \n\r <a href='https://digital.canada.ca'>Go back</a>",
-    "referrerUrlEn": "",
-    "referrerUrlFr": ""
-  },
-  "layout": [1, 2, 3, 4, 5, 7, 8, 9, 10, 12, 13, 14],
-  "elements": [
-    {
-      "id": 1,
-      "type": "richText",
-      "properties": {
-        "charLimit": 5000,
-        "validation": {
-          "required": false
-        },
-        "descriptionEn": "##Contact Information",
-        "descriptionFr": "##[fr] Contact Information"
-      }
-    },
-    {
-      "id": 2,
-      "type": "textField",
-      "properties": {
-        "titleEn": "1. What is your name?",
-        "titleFr": "[fr]1. What is your name?]",
-        "validation": {
-          "type": "text",
-          "required": true,
-          "maxLength": 150
-        },
-        "autoComplete": "name",
-        "descriptionEn": "",
-        "descriptionFr": "",
-        "placeholderEn": "",
-        "placeholderFr": ""
-      }
-    },
-    {
-      "id": 3,
-      "type": "radio",
-      "properties": {
-        "choices": [
-          {
-            "en": "English",
-            "fr": "Anglais"
-          },
-          {
-            "en": "French",
-            "fr": "Français"
-          }
-        ],
-        "titleEn": "2. What is your preferred language for communications?",
-        "titleFr": "[fr] 2. What is your preferred language for communications?",
-        "validation": {
-          "required": true
-        },
-        "descriptionEn": "",
-        "descriptionFr": ""
-      }
-    },
-    {
-      "id": 4,
-      "type": "richText",
-      "properties": {
-        "charLimit": 5000,
-        "validation": {
-          "required": false
-        },
-        "descriptionEn": "**Note:** We are not collecting email addresses or phone numbers through this test form to preserve your privacy \n\r ##Poject Information",
-        "descriptionFr": "[fr] **Note:** We are not collecting email addresses or phone numbers through this test form to preserve your privacy \n\r ##Project Information"
-      }
-    },
-    {
-      "id": 5,
-      "type": "textField",
-      "properties": {
-        "titleEn": "3. What is the name of your lemonade stand?",
-        "titleFr": "[fr]3. What is the name of your lemonade stand?",
-        "validation": {
-          "type": "text",
-          "required": true,
-          "maxLength": 300
-        },
-        "descriptionEn": "",
-        "descriptionFr": "",
-        "placeholderEn": "",
-        "placeholderFr": ""
-      }
-    },
-    {
-      "id": 7,
-      "type": "textArea",
-      "properties": {
-        "titleEn": "5. Please describe the taste of your lemonade recipe. ",
-        "titleFr": "[fr] 5. Please describe the taste of your lemonade recipe. ",
-        "validation": {
-          "type": "text",
-          "required": true,
-          "maxLength": 1000
-        },
-        "descriptionEn": "Consider including details about the sweetness or sourness of your lemonade recipe. Maximum characters: 1000",
-        "descriptionFr": "[fr]Consider including details about the sweetness or sourness of your lemonade recipe. Maximum characters: 1000",
-        "placeholderEn": "",
-        "placeholderFr": ""
-      }
-    },
-    {
-      "id": 8,
-      "type": "dropdown",
-      "properties": {
-        "choices": [
-          {
-            "en": "Park",
-            "fr": "[fr] Park"
-          },
-          {
-            "en": "Driveway",
-            "fr": "[fr]Driveway"
-          },
-          {
-            "en": "Building lobby",
-            "fr": "[fr]Building lobby"
-          },
-          {
-            "en": "Parking lot",
-            "fr": "[fr]Parking lot"
-          },
-          {
-            "en": "Other",
-            "fr": "[fr]Other"
-          }
-        ],
-        "titleEn": "6. Where is your lemonade stand going to be located?",
-        "titleFr": "[fr]6. Where is your lemonade stand going to be located?",
-        "validation": {
-          "required": true
-        },
-        "descriptionEn": "",
-        "descriptionFr": ""
-      }
-    },
-    {
-      "id": 9,
-      "type": "textField",
-      "properties": {
-        "titleEn": "7. If you checked Other, please specify",
-        "titleFr": "7. Pour d'autre information veuillez spécifier",
-        "validation": {
-          "type": "text",
-          "required": false,
-          "maxLength": 100
-        },
-        "descriptionEn": "",
-        "descriptionFr": "",
-        "placeholderEn": "",
-        "placeholderFr": ""
-      }
-    },
-    {
-      "id": 10,
-      "type": "richText",
-      "properties": {
-        "charLimit": 5000,
-        "validation": {
-          "required": false
-        },
-        "descriptionEn": "###Materials and Ingredients",
-        "descriptionFr": "###[fr] Materials and Ingredients"
-      }
-    },
-    {
-      "id": 12,
-      "type": "checkbox",
-      "properties": {
-        "choices": [
-          {
-            "en": "Cups",
-            "fr": "[fr] Cups"
-          },
-          {
-            "en": "Napkins",
-            "fr": "[fr]Napkins"
-          },
-          {
-            "en": "Straws",
-            "fr": "[fr]Straws"
-          }
-        ],
-        "titleEn": "9. Which materials will you purchase with this funding? ",
-        "titleFr": "[fr] 9. Which materials will you purchase with this funding? ",
-        "validation": {
-          "required": true
-        },
-        "descriptionEn": "",
-        "descriptionFr": ""
-      }
-    },
-    {
-      "id": 13,
-      "type": "richText",
-      "properties": {
-        "charLimit": 5000,
-        "validation": {
-          "required": false
-        },
-        "descriptionEn": "##Submission",
-        "descriptionFr": "##[fr] Submission"
-      }
-    },
-    {
-      "id": 14,
-      "type": "radio",
-      "properties": {
-        "choices": [
-          {
-            "en": "Yes",
-            "fr": "Oui"
-          },
-          {
-            "en": "No",
-            "fr": "Non"
-          }
-        ],
-        "titleEn": "10. Do you attest that the information you are providing is true and correct to your knowledge?",
-        "titleFr": "[fr] 10. Do you attest that the information you are providing is true and correct to your knowledge?",
-        "validation": {
-          "required": true
-        },
-        "descriptionEn": "",
-        "descriptionFr": ""
-      }
-    }
-  ]
-}
