--- components/clientComponents/form-builder/notes.md
@@ -1,8 +1,0 @@
-// Immer
-// https://www.youtube.com/watch?v=FmKjwh34Rn8
-
-// Zustand Re-renders
-// https://youtu.be/aOt4Hz3ze3Q?t=1547
-
-// Testing
-// https://docs.pmnd.rs/zustand/guides/testing
