--- components/clientComponents/form-builder/test-utils/index.ts
@@ -1,4 +1,0 @@
-import store from "./defaultStore.json";
-export const defaultStore = store;
-export { localStorageMock } from "./storage";
-export { Providers, withProviders } from "./Providers";
