--- components/clientComponents/form-builder/test-utils/storage.ts
@@ -1,18 +1,0 @@
-export const localStorageMock = (() => {
-  let store: Record<string, string> = {};
-
-  return {
-    getItem(key: string) {
-      return store[key] || null;
-    },
-    setItem(key: string, value: string) {
-      store[key] = value.toString();
-    },
-    removeItem(key: string) {
-      delete store[key];
-    },
-    clear() {
-      store = {};
-    },
-  };
-})();
