+++ lib/store/useTemplateStore.tsx
@@ -11,7 +11,7 @@ import {
 } from "zustand/middleware";
 
 import React, { createContext, useRef, useContext, useEffect } from "react";
-import { getPathString } from "../../components/clientComponents/form-builder/getPath";
+import { getPathString } from "../utils/form-builder/getPath";
 
 import {
   moveDown,
@@ -24,7 +24,7 @@ import {
   getSchemaFromState,
   incrementSubElementId,
   cleanInput,
-} from "../../components/clientComponents/form-builder/util";
+} from "../utils/form-builder";
 import { Language } from "../types/form-builder-types";
 import update from "lodash.set";
 import unset from "lodash.unset";
