+++ lib/utils/form-builder/__tests__/cleanInput.test.js
